add comments in RBT class
